requirements.txt
alembic==1.17.2
greenlet==3.2.4
Mako==1.3.10
MarkupSafe==3.0.3
pillow==12.0.0
psycopg2-binary==2.9.11
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
shiboken6==6.10.1
SQLAlchemy==2.0.44
tomli==2.3.0
typing_extensions==4.15.0
